Add thumbstick v2 design.
Flip the orientation of the model to put the mass into the top piece
instead. This requires using supports and some pegs to connect to the
bottom piece. Connecting this to the dpad and other pieces will allow us
to remove the need for supports. Pegs should be replaced with screwholes
and screws.