Add assembly and change select button in top plate.
Remove bottom plate to the select button holder in the top plate because
this will trap the switch when the wires are soldered. Instead make the
side pads longer and have them slot into the bottom plate.

Also add a screw hole near the select button switch to secure the select
button holder (to be made later).