Add bottom plate to left controller.
Complete the left controller frame assembly. The only thing missing now
is a little plate to hold the top of the select button in place.

NOTE to self: I've been thinking for a couple days that the extra
difficulty of creating a separate inner frame outweight the benefits of
having it when there aren't PCB's inside to hold everything in place. It
would probably be much stronger and easier to merge the top-plate into
the front face of the enclosure, separate the middle section that holds
the top of the dpad switches in place as its own piece, and separate the
shoulder button bottom section as its own piece.