Tweak left controller holder parts.
Increase the column height of the dpad holder fastener columns to 3.25mm
(+0.75mm) to allow for the dpad plastic cover to jiggle around more when
you press the buttons. This extra leeway feels much, much better.

Increase the column height of the select button holder by 1mm to 3mm
tall to allow for proper clearance of the squishy silicon membrane and
button. This is too high, but still works. Widen the button hole from
2.5mm radius to 2.6mm radius.

NOTE: the select button holder is flexible like paper and really shitty,
but it works for now. It's going to be revised out in the next iteration
anyway.