Add initial portion of the chassis.
This part contains screw holes and fastening surfaces for nuts on the
back of the screen and for screwing a protoboard in place. There is also
a hook jutting out for the left controller to fasten on to.

The printed result is pretty sturdy but has slight flex over all of the
part. The screen holes for nut fastening are a good thickness and size
to allow for the screws to go through them. They are all aligned. The
fastener columns for the right protoboard are in the right spot too. Reinforcements around the left controller hook might be necessary.
Overall, it looks really cool.

Need to add 3 fastener columns for the left protoboard as well as a hook
for the right controller frame. Also need to add two sections for
holding the speakers in place.