Make bottom plate changes on left controller frame.
Enlarge the dpad holder openings because even with 0.1mm clearance they
still got stuck. This is bad because when you go to remove the parts
from each other, the dpad holder legs can break off.

Change the shape of the chassis hook to be just a square. Need more
material around the screw holes.