Add ABXY jig.
This was surprisingly hard. It took a bit of iterating to find the right
spacing for the button holes and button-holder walls to hold the buttons
in place without having them spin around.

The vertical clearance of the button-holder and abxy-undercarriage was
also modified w.r.t. the dpad jig as well. The button holder top surface
is only 0.75mm thick vs 1mm thick of the dpad button holder. The columns
of both are 3mm total. The undercarriage of the abxy jig is not 1.85mm
like the dpad undercarriage, but has been expanded to be 2mm to allow
clearance for the things that jut out from underneath the tactile
switches. I suspect that the dpad undercarriage would actually benefit
from this too.