Fix right controller bottom plate alignment.
Fixed the alignment issues in the previous commit that were mentioned
about the right controller frame bottom plate. More specifically, the
breakout board bottom holder pad was moved 1mm to the left (the
guideline was incorrect) and the fastening hole was adjusted too. The
bottom holder pad was reduced in height by 1mm and moved upwards to
compensate. The shoulder buttons was also moved 1mm to the left due to
the same mistake as the previous change.