Fix right controller top/bottom plates.
Fixes to the right controller frame made after printing them out.

The top frame had the thumbstick wall cutouts too shallow. Changed these
to the right height. Also moved the thumbstick screw pads slightly to
make them fit better. This was not done to the original feature, but by
pocketing the old ones and then padding new ones in their place to avoid
breaking all the steps that happened after.

The bottom frame needed the thumbstick extrusion to be .75mm thinner on
the right and added a small reinforcement pad on the left. Decrease the
height of the undercarriage body by 0.15 to compensate from the columns
being 2mm instead of 1.85mm.

Add the prusa slicer file with all the parts to be printed laid out all
at once on the bed. Note that the shoulder buttons needed to be in the
middle because they are sensitive to warping if put near the edges or
something. The shoulder button design is sensitive to warped buttons and
will not work if the warping is too extreme. Also ironing is applied to
the shoulder buttons so that's captured by the slicer file.