Add WIP of screen-backing version 2.
Simplify the shapes and history tree of the screen-backing model. The
version 2 will have speaker holders, and changed from 2 protoboard
holders side by side to 1 protoboard holder on the right and a USB hub
holder on the right. The plan is to stack the two protoboards on top of
each other instead of side by side.

In this commit, the base shape is created. The features on the
controller hooks and the screw holes need to be made.