Bug fixes, add charging query, and hook up bq lib.
Fix memory leak related to the LCD display "de-syncing" or fogging
up when the script was exited. This happened because the
PNGVIEW_PROCESSES were not killed cleanly causing something to
crash. Cleaning up PNGVIEW_PROCESSES in on_exit() seems to fix
this.

Add bq27441 lib function to check the flags to see if the device
is discharging. Use the boolean negation of this function to see
if the device is charging.

Bug fix related to the above where the charging variable was not
correctly routed through draw_hud() and the battery image drawing
function.

Hook up the bq27441 lib function call to the display code logic
through the SOC interrupt.