Changes made to second tier holder in chassis-v2
Thicken the USB Holder Wall to match the first version. This was
accidentally made too thin in the second version. This necessitated
moving the audio wires hole a little bit to the side. This should
have no impact on the alignment of other parts.

Remove one screw hole under the power switch footprint and add a
pin clearance cutout hole instead.

Add a pin clearance cutout hole to the 5V regulator footprint.