Changes to second tier holder v2 in chassis-v2.
Remove USB cable press. Instead create a pocket to give the cable
extra clearance.

Also fix many other issues:

Widen the screw access holes in the controller button pads so that
the 1.6M screws with a 1mm tall head can fit in it.

Center the power switch holder screw holes in the middle of the pad.